[COOK-3029] - Add recipe for managing Oracle Java RPM.
oracle_rpm recipe assumes the Oracle java RPM is available in a custom
YUM repositories.

This recipe is configured with the following attributes:

      default['java']['install_flavor'] = 'oracle_rpm'
      default['java']['oracle_rpm']['type'] = 'jdk'

The type attribute is either "jdk" or "jre".

This recipe also sets the JAVA_HOME environment variable and enables
the commands included in the RPM with the update-alternatives.