Download java windows binaries from S3 with creds
Its useful to stuff the java blob into S3 to easily install it across
your own servers, but making this --acl-public on S3 will violate the
oracle license.  This patch is a little messy but accomplishes that
goal.  I didn't like the approach of trying to shove it into the
windows_package provider.  I also don't like pushing the S3 creds into
the node attributes rather than using encrypted databags, but I think
we need to do that for now otherwise you wind up with madness where
you have all kinds of node attrs to select what kind of download you
are doing.

The right way to fix this is through being able to inject information
into cookbooks dynamically:

https://github.com/danielsdeleo/chef-data-bindings

Since that is the right direction to go and any direction we take here
is going to be throwaway work using the wrong primitives, I didn't
try to make this perfect.  We can point to this code, instead, as
a reason why we need something like the chef data bindings.

(And I need to ship this to support the opscode-ci cookbook and stop
that train and allow the perfect to be the enemy of the good...)