[COOK-3704] Fix alternatives when the package is already installed
The current Java cookbook uses a conditional notification on the package installation to notify the java alternatives to update. This actually makes it impossible to switch between java versions, because the default java being set is dependent on the package installation, not the value of the alternative.

This patch removes the notification and checks the current value of the alternative to see if it should be updated. This ensure that the correct Java is used, even if the package was already installed.