Moving configuration dependent on version/install_flavor to a recipe
This is so that the user can override node['java']['jdk_version'] and
node['java']['install_flavor'] before they are used to compute other
attributes, such as openjdk_packages. This is necessary to make it
possible to override version/install_flavor in a data bag using the
bag_config cookbook.