Fix JDK install on Windows.
Commit b42c246d introduces quotes around the install path, so that
we can safely provide paths with strings. Unfortunately, it seems
that this method works only with JREs, not with JDKs.
It seems that the recommended way to call the installer (described at
http://bugs.sun.com/bugdatabase/view_bug.do?bug_id=4966488)
is by adding /v"/qn INSTALLDIR=\"path\""
I tested that this change makes it possible to install both a JRE
and a JDK.