Change ownership of extracted files
Change ownership of all extracted JDK files to the user and group as
specified by the attribute 'owner'. Otherwise the ownership of these
files is set to some random UID/GID as found in original archive (e.g.
all files in jdk-7u55-linux-x64.tar are set to UID=10 and GID=143).

Signed-off-by: Gregor Zurowski <gregor@zurowski.org>