Bug Fix: tar exit with error status 2 when extract java tarball to chef cache folder which is mounted on NFS
The error comes from that the guest machine is not able to set the owner of the files.  Since the /var/chef/cache folder is mounted via NFS, even root had problem setting the correct permissions and therefore failed. Adding --no-same-owner parameter to tar command would avoid the failure. And it is OK becasue chmod is called when extracted folder is moved to correct place