Set connect timeout to a rationally short value
This seems like it's being treated as a download time limit, but "connect timeout" is really a limit on establishing a connection to the server.  Typically if it doesn't happen in ~30s, it's not going to happen at all.

Don't make convergences wait 10m for a connection that isn't going to happen.