Use stable ChefDK in Travis and skip chef-sugar
When we promote from unstable to stable at the moment we do a move which means for some period of time there is no build in the current channel. This causes travis runs to fail so we recommend not using current in travis for now. We’re hoping to update how promotion occurs soon, which would allow for the use of current in travis once again.

Signed-off-by: Tim Smith <tsmith@chef.io>