node `bin_cmds` attr not properly set for ibm java
When the `install_flavor` attribute is set to `ibm`, the attributes for
the JDK 6, 7, and 8 `bin_cmds` were not being set. This was caused by an
issue in attribute precedence.

In the `attributes/default.rb` file, the `bin_cmds` attributes were set
based on a case statement which was triggered by the `install_flavor`.
By default, the install flavor is set to `openjdk`. Based on how chef
behaves during the compilation phase, the case statement is triggered
when this attributes file is loaded which means the value of
`install_flavor` will *always* be `openjdk` at that moment so the
conditional branch for `ibm` is never invoked and thus the `bin_cmds`
attributes aren't set. Even if users `force_default` or `override` the
`install_flavor` to be `ibm` in their own cookbook.

This bug was causing the error observed in issue #271.

The fix removes the case statement completely. Since the attributes for
IBM and Oracle RPM are well scoped under `node['java']['ibm']` and
`node['java']['oracle_rpm']`